+++ src/RegisterGroup.h
@@ -45,7 +45,7 @@ struct GroupUtility
      * @param string claim - indicates the data which mapping requires
      * @return the address as int
      */
-    static uint16_t getAddressFromString(const std::string& string);
+    static int32_t getAddressFromString(const std::string& string);
 
     /**
      * @brief Utility method to fetch the bit from claim string
@@ -109,13 +109,13 @@ public:
 
     void setDevice(const std::shared_ptr<ModbusDevice>& device);
 
-    uint16_t getStartingAddress() const;
+    int32_t getStartingAddress() const;
 
     uint16_t getAddressCount() const;
 
-    int8_t getSlaveAddress() const;
+    int16_t getSlaveAddress() const;
 
-    void setSlaveAddress(int8_t slaveAddress);
+    void setSlaveAddress(int16_t slaveAddress);
 
     bool isReadRestricted() const;
 
@@ -142,7 +142,7 @@ private:
     bool keyExistsInSet(const std::string& key);
 
     RegisterMapping::RegisterType m_registerType;
-    int8_t m_slaveAddress;
+    int16_t m_slaveAddress;
     bool m_readRestricted;
 
     std::shared_ptr<ModbusDevice> m_device;
