+++ src/mappings/FloatMapping.h
@@ -38,8 +38,8 @@ public:
      * @param readRestricted Is the Mapping write only?
      * @param slaveAddress Slave address of device, leave to be assigned by device, default is -1.
      */
-    FloatMapping(const std::string& reference, RegisterType registerType, const std::vector<int16_t>& addresses,
-                 bool readRestricted = false, int8_t slaveAddress = -1);
+    FloatMapping(const std::string& reference, RegisterType registerType, const std::vector<int32_t>& addresses,
+                 bool readRestricted = false, int16_t slaveAddress = -1);
 
     /**
      * @details Override methods will be executed on devices reading thread, so that this parsing can be done
