+++ src/mappings/Int16Mapping.h
@@ -38,8 +38,8 @@ public:
      * @param readRestricted Is the Mapping write only?
      * @param slaveAddress Slave address of device, leave to be assigned by device, default is -1.
      */
-    Int16Mapping(const std::string& reference, RegisterType registerType, int16_t address, bool readRestricted = false,
-                 int8_t slaveAddress = -1);
+    Int16Mapping(const std::string& reference, RegisterType registerType, int32_t address, bool readRestricted = false,
+                 int16_t slaveAddress = -1);
 
     /**
      * @details Override methods will be executed on devices reading thread, so that this parsing can be done
