+++ src/mappings/Int32Mapping.cpp
@@ -25,8 +25,8 @@
 namespace wolkabout
 {
 Int32Mapping::Int32Mapping(const std::string& reference, RegisterMapping::RegisterType registerType,
-                           const std::vector<int16_t>& addresses, RegisterMapping::OperationType operation,
-                           bool readRestricted, int8_t slaveAddress)
+                           const std::vector<int32_t>& addresses, RegisterMapping::OperationType operation,
+                           bool readRestricted, int16_t slaveAddress)
 : RegisterMapping(reference, registerType, addresses, OutputType::INT32, operation, readRestricted, slaveAddress)
 {
     if (operation != OperationType::MERGE_BIG_ENDIAN && operation != OperationType::MERGE_LITTLE_ENDIAN)
