+++ src/mappings/StringMapping.cpp
@@ -25,8 +25,8 @@
 namespace wolkabout
 {
 StringMapping::StringMapping(const std::string& reference, RegisterMapping::RegisterType registerType,
-                             const std::vector<int16_t>& addresses, RegisterMapping::OperationType operation,
-                             bool readRestricted, int8_t slaveAddress)
+                             const std::vector<int32_t>& addresses, RegisterMapping::OperationType operation,
+                             bool readRestricted, int16_t slaveAddress)
 : RegisterMapping(reference, registerType, addresses, OutputType::STRING, operation, readRestricted, slaveAddress)
 {
     if (operation != OperationType::STRINGIFY_ASCII && operation != OperationType::STRINGIFY_UNICODE)
