+++ src/mappings/UInt16Mapping.cpp
@@ -22,8 +22,8 @@
 
 namespace wolkabout
 {
-UInt16Mapping::UInt16Mapping(const std::string& reference, RegisterMapping::RegisterType registerType, int16_t address,
-                             bool readRestricted, int8_t slaveAddress)
+UInt16Mapping::UInt16Mapping(const std::string& reference, RegisterMapping::RegisterType registerType, int32_t address,
+                             bool readRestricted, int16_t slaveAddress)
 : RegisterMapping(reference, registerType, address, readRestricted, slaveAddress)
 {
     if (!(registerType == RegisterType::INPUT_REGISTER || registerType == RegisterType::HOLDING_REGISTER))
